+++ Module/en-US/SplitPipeline.dll-Help.ps1
@@ -14,14 +14,18 @@ Import-Module SplitPipeline
 	Splits pipeline input and processes input parts by parallel pipelines.
 '@
 	description = @'
-	The cmdlet splits pipeline input in real time and processes input parts by
-	parallel pipelines. The algorithm works without having the entire input
+	The cmdlet splits pipeline input and processes input parts by parallel
+	pipelines. The algorithm starts to work without having the entire input
 	available. Input can be very large or even infinite.
 
-	The cmdlet creates several pipelines. Each new pipeline is created as soon
-	as input is available, existing pipelines are busy, and number of created
-	pipelines is less than Count. Input items are processed by parts, part
-	sizes are defined by the Load parameter but not necessarily equal to it.
+	Input is processed by parts. Size of each part is defined by the parameter
+	Load but not necessarily equal to it. The Limit is used in order to define
+	the maximum part size.
+
+	The cmdlet creates several pipelines. Each pipeline is created when input
+	parts are available, created pipelines are busy, and their number is less
+	than Count. Created pipelines are used for processing several input parts,
+	one at a time.
 
 	The Begin and End script are invoked for each created pipeline before and
 	after processing. Each input part is piped to the script Script which is
@@ -31,17 +35,14 @@ Import-Module SplitPipeline
 	then incoming input items are enqueued for later processing. If the queue
 	size hits the limit Queue then the algorithm waits for a ready pipeline.
 
-	Input items are not necessarily processed in the same order as they come
-	into the cmdlet. But output can be ordered according to input, use Order.
-
-	The cmdlet is not recommended for scenarios with slow input, that is when
-	input items arrive slower than they are processed by the specified script.
+	Input items are not necessarily processed in the same order as they come.
+	But output can be ordered according to input parts, use the switch Order.
 '@
 	parameters = @{
 		Script = @'
 		The script invoked for each input part of each pipeline with an input
 		part piped to it. The script either processes the whole part ($input)
-		or each item ($_) separately in the process block. Examples:
+		or each item ($_) separately in the "process" block. Examples:
 
 			# Process a whole $input part:
 			... | Split-Pipeline { $input | %{ $_ } }
@@ -49,22 +50,23 @@ Import-Module SplitPipeline
 			# Process each item $_ separately:
 			... | Split-Pipeline { process { $_ } }
 
-		A script may have all three special blocks begin/process/end:
+		A script may have any of special blocks "begin", "process", and "end":
 
 			... | Split-Pipeline { begin {...} process { $_ } end {...} }
 
-		Note that these begin and end blocks are called for each input part,
-		unlike scripts Begin and End (parameters) called for each pipeline.
+		Note that such "begin" and "end" blocks are called for input parts but
+		scripts defined by parameters Begin and End are called for pipelines.
 '@
 		Begin = @'
-		The script invoked for each pipeline once before processing. The goal
-		of this script is to initialize the current runspace, normally to set
-		some variables, dot-source scripts, import modules, and etc.
+		The script invoked for each pipeline on creation before processing. The
+		goal is to initialize the runspace to be used by the pipeline, normally
+		to set some variables, dot-source scripts, import modules, and etc.
 '@
 		End = @'
 		The script invoked for each pipeline once after processing. The goal
 		is, for example, to output some results accumulated during processing
-		of all input parts.
+		of input parts by the pipeline. Consider to use Finally for releasing
+		resources instead of End or in addition to it.
 '@
 		Finally = @'
 		The script invoked for each pipeline before its closing. The goal is to
@@ -76,8 +78,9 @@ Import-Module SplitPipeline
 		Count = @'
 		Maximum number of created parallel pipelines. The default value is the
 		number or processors. Use the default or even decrease it for intensive
-		jobs. But for jobs not consuming much processor resources increasing
-		this number may improve overall performance a lot.
+		jobs, especially if there are other tasks working at the same time, for
+		example, output is processed simultaneously. But for jobs not consuming
+		much processor resources increasing the number may improve performance.
 '@
 		Queue = @'
 		Maximum number of objects in the queue. If all pipelines are busy then
@@ -85,16 +88,19 @@ Import-Module SplitPipeline
 		unlimited by default. The limit should be specified for potentially
 		large input. When the limit is hit the engine waits for a pipeline
 		available for input from the queue.
+
+		CAUTION: The Queue limit may be ignored and exceeded if Refill is used.
+		Any number of objects written via [ref] go straight to the input queue.
 '@
 		Auto = @'
 		Tells to tune some parameters automatically during processing in order
 		to increase utilization of pipelines and reduce overhead. This is done
-		normally by increasing the value of Load. Use Verbose in order to show
+		normally by increasing the value of Load. Use Verbose in order to view
 		some details during and after processing.
 
 		Note that using of a reasonable initial Load value known from practice
 		may be still useful with Auto, the algorithm may work effectively from
-		the start and still be able to adjust the Load dynamically.
+		the start and still be able to adapts the Load dynamically.
 
 		Use Cost in order to specify internal overhead range. This is unlikely
 		needed in most cases, the default range may be just fine.
@@ -103,7 +109,8 @@ Import-Module SplitPipeline
 		Recommended percentage of inner overhead time with respect to overall
 		time. It is used together with the switch Auto and ignored otherwise.
 		It accepts one or two values: the maximum (first) and minimum (second)
-		percentage. Values 5, 1 are used by default (may change in vNext).
+		percentage. If the second value is omitted then 0 is assumed. Default
+		values are 5 and 1 (they may change in future versions).
 '@
 		Load = @'
 		Recommended minimum number of input objects for each parallel pipeline.
@@ -111,14 +118,14 @@ Import-Module SplitPipeline
 		improve overall performance, it may reduce the total number of input
 		parts and overhead of pipeline invocations for each part.
 
-		Use the switch Auto in order to define the Load during processing. But
-		a proper initial value known from practice may be still important.
+		Use the switch Auto in order to tune the Load during processing. But a
+		proper initial value known from practice may be still useful with Auto.
 '@
 		Limit = @'
 		Maximum number of input objects for each parallel pipeline. The default
 		is 0 (unlimited). Setting this limit causes more frequent output (if it
 		is actually hit). This may be important for feeding downstream commands
-		working at the same time.
+		in the pipeline working at the same time.
 '@
 		Variable = @'
 		Variables imported from the current runspace to parallel.
@@ -131,7 +138,13 @@ Import-Module SplitPipeline
 '@
 		Order = @'
 		Tells to output part results in the same order as input parts arrive.
-		The algorithm may work slightly slower.
+		The algorithm may work a little bit slower.
+'@
+		Refill = @'
+		Tells to refill the input by [ref] objects from output. Other objects
+		go to output as usual. This convention is used for processing items of
+		hierarchical data structures: child container items come back to input,
+		leaf items or other data produced by processing go to output.
 '@
 		InputObject = @'
 		Input objects processed by parallel pipelines. Do not use this
@@ -166,6 +179,32 @@ Import-Module SplitPipeline
 	Two commands perform the same job simulating long but not processor
 	consuming operations of each item. The first command takes about 10
 	seconds. The second takes about 2 seconds due to Split-Pipeline.
+'@
+			test = { . $args[0] }
+		}
+		@{
+			code = {
+				$PSHOME | Split-Pipeline -Refill {process{
+					foreach($item in Get-ChildItem -LiteralPath $_ -Force) {
+						if ($item.PSIsContainer) {
+							[ref]$item.FullName
+						}
+						else {
+							$item.Length
+						}
+					}
+				}} | Measure-Object -Sum
+			}
+			remarks = @'
+	This is an example of Split-Pipeline with refilled input. The convention:
+	[ref] objects refill the input, other objects go to output as usual.
+
+	The code calculates the number and size of files in $PSHOME. It is a "how
+	to" sample, performance gain is not expected because the code is trivial
+	and works relatively fast.
+
+	See also another example with simulated slow data requests:
+	https://github.com/nightroman/SplitPipeline/blob/master/Tests/Test-Refill.ps1
 '@
 			test = { . $args[0] }
 		}